Mouse Aiming & Shoot function
